requirements.txt
fastapi
uvicorn
python-multipart
jinja2
torch
torchvision
numpy
pillow
scikit-learn
